correct program

+++ lab4/func.h
@@ -12,7 +12,7 @@ int def_len(unsigned char type_h, unsigned char type_c, char *hash, char *cipher
 int check(char *code, char hash, char cipher);
 int compare(unsigned char *s1);
 unsigned char *ASCII_to_hhx(char *s, unsigned char *x);
-unsigned char *int_to_hhx(int in, unsigned char *x);
+unsigned char *int_to_hhx(unsigned long int in, unsigned char *x);
 unsigned char *gen_text(unsigned char *text, size_t len);
 unsigned char *hmac_md5(unsigned char *nonce, size_t nonce_len, unsigned char *password, size_t password_len, unsigned char *key);
 unsigned char *hmac_sha1(unsigned char *nonce, size_t nonce_len, unsigned char *password, size_t password_len, unsigned char *key);
